[CODEMOVE] moving renderer.ts to renderer folder
1. moving renderer.ts to renderer folder
2. also moving the channel_logger.ts and process_folder.ts to lib folder
