+++ lib/channel_logger.ts
@@ -1,5 +1,4 @@
 import { BrowserWindow } from "electron";
-import log from "electron-log";
 
 
 /**
