+++ lib/process_folder.ts
